refactor: reorganize to standard Python package structure
Major changes:
- Create src/mcp_image_server/ package with providers/ and transports/ modules
- Move all tests to tests/, examples to examples/, docs to docs/
- Add entry points: mcp-server script, python -m mcp_image_server
- Update all imports to use relative imports within package
- Update README with new commands and structure
