--- README_UPDATE_SUMMARY.md
@@ -1,174 +1,0 @@
-# README更新总结
-
-## 📋 更新概述
-
-README文件已成功更新，新增了多API支持的完整说明，同时保持了原有的图片和架构内容。
-
-## 🔄 主要更新内容
-
-### 1. 项目描述更新
-- **原来**: 仅支持腾讯混元API的图像生成服务
-- **现在**: 支持多个主流AI提供商（腾讯混元、OpenAI DALL-E 3、豆包）的统一服务
-
-### 2. 新增功能特性说明
-
-#### 🎯 多API提供商支持
-- **腾讯混元**: 18种艺术风格，中文优化
-- **OpenAI DALL-E 3**: 高质量图像生成，英文优化  
-- **豆包（字节跳动）**: 平衡的质量和速度，12种风格
-
-#### 🔧 智能提供商管理
-- 自动检测可用的API提供商
-- 支持指定特定提供商或自动选择
-- 统一的错误处理和重试机制
-- 灵活的参数格式：`provider:style` 和 `provider:resolution`
-
-### 3. 使用方法更新
-
-#### 服务器版本选择
-- 新增了两种服务器实现的说明：
-  - **单API服务器**（原版）：`mcp_image_server.py`
-  - **多API服务器**（新版，推荐）：`mcp_image_server_multi.py`
-
-#### 多API使用示例
-添加了详细的使用示例：
-```python
-# 自动选择最佳提供商
-generate_image(prompt="A cute cat in a garden")
-
-# 指定特定提供商
-generate_image(prompt="A cute cat", provider="openai")
-
-# 使用提供商特定的风格和分辨率
-generate_image(
-    prompt="Cyberpunk city skyline", 
-    style="hunyuan:saibopengke", 
-    resolution="hunyuan:1024:768"
-)
-```
-
-### 4. 支持的提供商和参数
-
-新增了完整的提供商参数对照表：
-
-| 提供商 | 风格数量 | 分辨率选项 | 特色功能 |
-|--------|----------|------------|----------|
-| **腾讯混元** | 18种 | 8种 | 中文优化，艺术风格丰富 |
-| **OpenAI DALL-E 3** | 12种 | 7种 | 高质量，英文优化 |
-| **豆包** | 12种 | 9种 | 平衡质量与速度 |
-
-### 5. Cursor集成配置更新
-
-#### 新增多API配置示例
-```json
-{
-  "mcpServers": {
-    "multi-image-generation": {
-      "name": "Multi-API Image Generation Service",
-      "description": "Multi-provider image generation using Hunyuan, OpenAI, and Doubao APIs",
-      "type": "stdio",
-      "command": "path/to/python.exe",
-      "args": ["path/to/mcp_image_server_multi.py"],
-      "environment": [
-        "TENCENT_SECRET_ID", 
-        "TENCENT_SECRET_KEY",
-        "OPENAI_API_KEY",
-        "DOUBAO_ACCESS_KEY",
-        "DOUBAO_SECRET_KEY",
-        "MCP_IMAGE_SAVE_DIR"
-      ]
-    }
-  }
-}
-```
-
-#### 环境变量配置说明
-- 为单API和多API配置分别提供了详细的环境变量说明
-- 明确指出用户只需配置想要使用的提供商的API密钥
-
-### 6. API参考文档更新
-
-#### 新增多API架构说明
-- **支持的API**: 腾讯混元、OpenAI DALL-E 3、豆包
-- **统一MCP资源**: `providers://list`、`styles://list`、`resolutions://list`
-- **增强的MCP工具**: 多提供商图像生成与智能路由
-
-#### 各API提供商详细说明
-为每个新增的API提供商添加了详细的特性说明：
-- OpenAI DALL-E 3 API特性
-- 豆包API（字节跳动）特性
-
-### 7. RoadMap更新
-
-#### 当前版本成就
-- ✅ 腾讯混元图像生成API
-- ✅ OpenAI DALL-E 3 API集成
-- ✅ 豆包API集成
-- ✅ 多提供商管理系统
-- ✅ 智能提供商选择
-- ✅ 统一参数接口
-
-#### 未来计划扩展
-- 支持更多主流API（阿里通义万相、百度文心一格、Stable Diffusion）
-- 高级功能（图像编辑、批量生成、风格转换）
-- 增强MCP集成（实时进度、图像历史管理）
-
-## 🖼️ 保持的原有内容
-
-### 完整保留的图片资源
-- MCP服务器运行截图
-- Cursor集成步骤截图
-- 图片生成和移动演示
-- 前端演示网站截图
-- Windsurf IDE集成截图
-
-### 保留的架构内容
-- 安装说明（UV和传统pip）
-- 环境变量配置
-- 验证步骤
-- 故障排除指南
-- 前端演示说明
-- 兼容性说明
-- 致谢和贡献指南
-
-## 🌐 多语言支持
-
-同时更新了中英文README：
-- **README.md**: 英文版本，包含完整的多API支持说明
-- **README.zh-CN.md**: 中文版本，对应的中文说明
-
-## ✅ 更新验证
-
-### 内容完整性检查
-- ✅ 所有原有图片链接保持完整
-- ✅ 原有架构说明完全保留
-- ✅ 新增API支持说明详细完整
-- ✅ 配置示例准确可用
-- ✅ 使用示例清晰易懂
-
-### 文档结构优化
-- ✅ 逻辑结构清晰，从基础到高级
-- ✅ 新旧功能对比明确
-- ✅ 配置选项分类清楚
-- ✅ 代码示例格式规范
-
-## 🎯 用户受益
-
-通过此次README更新，用户可以：
-
-1. **清楚了解项目演进**: 从单API到多API的发展历程
-2. **快速选择适合的版本**: 单API vs 多API服务器
-3. **轻松配置多提供商**: 详细的配置指南和示例
-4. **灵活使用不同API**: 丰富的使用示例和参数说明
-5. **无缝迁移升级**: 向后兼容的升级路径
-
-## 📝 总结
-
-README更新成功完成了以下目标：
-- ✅ **保持原有内容**: 所有图片和架构内容完整保留
-- ✅ **新增API支持说明**: 详细的多API功能介绍
-- ✅ **提供完整配置指南**: 从安装到使用的全流程说明
-- ✅ **确保向后兼容**: 原有功能继续可用
-- ✅ **优化用户体验**: 清晰的结构和丰富的示例
-
-项目README现在提供了完整、准确、易用的多API图像生成服务文档，为用户提供了从入门到高级使用的全方位指导。
