+++ examples/image_generation_tool.py
