+++ mcp-server
@@ -1,0 +1,10 @@
+#!/usr/bin/env python3
+"""
+Entry point script for MCP Image Generation Server.
+
+This script can be used to start the server directly from the project root.
+"""
+
+if __name__ == "__main__":
+    from mcp_image_server.main import main
+    main()
