+++ src/mcp_image_server/__init__.py
@@ -1,0 +1,10 @@
+"""
+MCP Image Generation Server
+
+A Model Context Protocol (MCP) server that provides image generation capabilities
+through multiple API providers (Hunyuan, OpenAI, Doubao).
+
+Supports both stdio and HTTP transports.
+"""
+
+__version__ = "0.2.0"
