+++ src/mcp_image_server/__main__.py
@@ -1,0 +1,8 @@
+"""
+Allow the package to be run as a module: python -m mcp_image_server
+"""
+
+from .main import main
+
+if __name__ == "__main__":
+    main()
