+++ src/mcp_image_server/providers/__init__.py
