+++ src/mcp_image_server/providers/base.py
