+++ src/mcp_image_server/providers/doubao_provider.py
