+++ src/mcp_image_server/providers/hunyuan_provider.py
