+++ src/mcp_image_server/providers/openai_provider.py
