+++ src/mcp_image_server/providers/provider_manager.py
