+++ src/mcp_image_server/transports/__init__.py
@@ -1,0 +1,12 @@
+"""
+Transport layer modules for MCP Image Generation Server.
+
+Supports:
+- stdio: Standard input/output transport for local IDE integration
+- HTTP: HTTP transport for remote access with session management
+"""
+
+from .stdio import *
+from .http_server import *
+
+__all__ = ["stdio", "http_server", "http", "auth", "session_manager"]
