+++ src/mcp_image_server/transports/auth.py
