+++ src/mcp_image_server/transports/session_manager.py
