+++ tests/test_stdio_quick.py
