+++ tests/test_stdio_simple.py
