+++ tests/test_stdio_transport.py
